moved tests from LinearAlgebra to Point4
